+++ crates/teloxide-core/src/types/inline_query_result_cached_mpeg4_gif.rs
@@ -37,6 +37,10 @@ pub struct InlineQueryResultCachedMpeg4Gif {
     /// specified instead of `parse_mode`.
     pub caption_entities: Option<Vec<MessageEntity>>,
 
+    /// Pass `true`, if the caption must be shown above the message media.
+    #[serde(default, skip_serializing_if = "std::ops::Not::not")]
+    pub show_caption_above_media: bool,
+
     /// [Inline keyboard] attached to the message.
     ///
     /// [Inline keyboard]: https://core.telegram.org/bots#inline-keyboards-and-on-the-fly-updating
@@ -60,6 +64,7 @@ impl InlineQueryResultCachedMpeg4Gif {
             parse_mode: None,
             reply_markup: None,
             caption_entities: None,
+            show_caption_above_media: false,
             input_message_content: None,
         }
     }
@@ -99,6 +104,11 @@ impl InlineQueryResultCachedMpeg4Gif {
         C: IntoIterator<Item = MessageEntity>,
     {
         self.caption_entities = Some(val.into_iter().collect());
+        self
+    }
+
+    pub fn show_caption_above_media(mut self, val: bool) -> Self {
+        self.show_caption_above_media = val;
         self
     }