+++ crates/teloxide-core/src/types/inline_query_result_video.rs
@@ -45,6 +45,10 @@ pub struct InlineQueryResultVideo {
     /// specified instead of `parse_mode`.
     pub caption_entities: Option<Vec<MessageEntity>>,
 
+    /// Pass `true`, if the caption must be shown above the message media.
+    #[serde(default, skip_serializing_if = "std::ops::Not::not")]
+    pub show_caption_above_media: bool,
+
     /// Video width.
     pub video_width: Option<u32>,
 
@@ -92,6 +96,7 @@ impl InlineQueryResultVideo {
             caption: None,
             parse_mode: None,
             caption_entities: None,
+            show_caption_above_media: false,
             video_width: None,
             video_height: None,
             video_duration: None,
@@ -154,6 +159,11 @@ impl InlineQueryResultVideo {
         C: IntoIterator<Item = MessageEntity>,
     {
         self.caption_entities = Some(val.into_iter().collect());
+        self
+    }
+
+    pub fn show_caption_above_media(mut self, val: bool) -> Self {
+        self.show_caption_above_media = val;
         self
     }